✨ Leverage entity unit when relevant
Before this patch, all metrics where unitless. It made dashboards in
datadog hearder to read (without any unit).

Now we use specific metric name when the state has a unit. It allows to
set the unit on the metric metadata in datadog and thus have more
readable dashboard.

💥 Sadly it means we break existing dashboards.
Because the integration is still in development, it's probably
acceptable.