✨ Add entity/device/area labels as tags
This PR adds tags based on entity relationship with device and areas by
using their labels as tags.

For instance, brightness entity of a light located in a bedroom will
have tags based on the labels on the entity, its device and the area.

This should help with building monitors/dashboards based on properties
instead of listing explicitely entity names.